+++ database/migrations/20200419103403_add_external_services_to_sessions_table.php
@@ -1,0 +1,21 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class AddExternalServicesToSessionsTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('sessions', function (Blueprint $table) {
+            $table->text('externalservices')->nullable();
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('sessions', function (Blueprint $table) {
+            $table->dropColumn('externalservices');
+        });
+    }
+}